Reduxの環境構築
